Installation mariadb

+++ server/database/models/Ajustement_detail.model.js
@@ -4,8 +4,8 @@ import db from "../../config/Database.js";
 const Ajustement_detail = db.define(
   "ajustement_detail",
   {
-    qantite_nouveau: { type: DataTypes.DOUBLE(10), allowNull: false },
-    qantite_ancien: { type: DataTypes.DOUBLE(10), allowNull: false },
+    qantite_nouveau: { type: DataTypes.DOUBLE, allowNull: false },
+    qantite_ancien: { type: DataTypes.DOUBLE, allowNull: false },
     status: { type: DataTypes.BOOLEAN, defaultValue: "1" },
   },
   { paranoid: true, freezeTableName: true }
